requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
imutils==0.5.4
joblib==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.0
opencv-contrib-python==4.5.5.62
opencv-python-headless==4.5.5.62
Pillow==8.2.0
protobuf==3.19.4
psycopg2==2.9.3
pytz==2021.3
regex==2022.3.15
requests==2.27.1
scikit-image==0.19.1
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
threadpoolctl==3.1.0
typing_extensions==4.0.1
whitenoise==5.3.0
